requirements.txt
python-dotenv
dotenv
fastapi[standard]
python-multipart
uvicorn
google-generativeai
pdfplumber